+++ R/prepare.R
@@ -284,6 +284,7 @@ prepare_ceteris_paribus <- function(x, variables = NULL) {
 }
 
 prepare_feature_importance <- function(x, max_features = 10, show_boxplot = TRUE,
+                                       x_title = "drop-out loss",
                                        margin = 0.15, digits = 3, rounding_function = round) {
   ### This function returns object needed to plot FeatureImportance in D3 ###
 
@@ -352,6 +353,7 @@ prepare_feature_importance <- function(x, max_features = 10, show_boxplot = TRUE
   ret$x <- new_x
   ret$m <- m
   ret$x_min_max <- min_max
+  ret$x_title <- x_title
   ret$desc <- data.frame(type = "desc",
                          text = gsub("\n","</br>", desc))